chore: add logs to pre and post plugins

+++ packages/autorelease-task-write-changelog/src/index.js
@@ -6,19 +6,22 @@ const readFile = promisify(_readFile);
 const writeFile = promisify(_writeFile);
 
 export default async function(ctx) {
-	const {changelog,basedir=".",options={}} = ctx;
+	const {changelog,basedir=".",options={},dryrun} = ctx;
 	const {changelogFile="changelog.md"} = options;
 
 	if (typeof changelog !== "string") {
 		return;
 	}
 
-	// write the changelog to a file
 	const cfile = resolve(basedir, changelogFile);
-	let existing = "";
-	try { existing = await readFile(cfile, "utf-8"); }
-	catch(e) { if (e.code !== "ENOENT") throw e; }
-	await writeFile(cfile, changelog + existing);
+
+	// prepend the changelog to file
+	if (!dryrun) {
+		let existing = "";
+		try { existing = await readFile(cfile, "utf-8"); }
+		catch(e) { if (e.code !== "ENOENT") throw e; }
+		await writeFile(cfile, changelog + existing);
+	}
 
 	return cfile;
 }
